func(getter): getPageDataColumn to access pages in cells
This function was created mainly for use with displaying
the table. It can use the for loop counter as a parameter
instead of doing row calculations manually.

This function accesses sections of pageData as if it was
a table. But instead of accessing it by row and column like:

/*If the matrix/table is 3x3*/
(0,0),(0,1),(0,2),(1,0),(1,1),(1,2),(2,0),(2,1),(2,2)

It accesses them like:
0:(0,0) , 1(0,1), 2(0,2), 3(1,0), 4(1,1)...